+++ src/analyzers/concept/freelancer-rates/index.ts
@@ -1,0 +1,5 @@
+import { ExemplarAnalyzer } from '../__exemplar'
+
+export class FreelancerRatesAnalyzer extends ExemplarAnalyzer {
+  // TODO: implement actual analyzer
+}